+++ src/server/api/endpoints/admin/delete-all-files-of-a-user.ts
@@ -1,7 +1,7 @@
 import $ from 'cafy';
 import define from '../../define';
 import { deleteFile } from '../../../../services/drive/delete-file';
-import { DriveFiles } from '../../../../models';
+import { DriveFiles, Users } from '../../../../models';
 import { ID } from '../../../../misc/cafy-id';
 
 export const meta = {
@@ -25,6 +25,16 @@ export default define(meta, async (ps, me) => {
 	const files = await DriveFiles.find({
 		userId: ps.userId
 	});
+
+	const user = await Users.findOne(ps.userId as string);
+
+	if (user.isAdmin) {
+		throw new Error('cannot delete files of admin');
+	}
+
+	if (me.isModerator && user.isModerator) {
+		throw new Error('cannot delete files of moderator');
+	}
 
 	for (const file of files) {
 		deleteFile(file);
